Fix processMessages such that MessageResults are returned in a deque
The previous implementation would only return the last message rather than
all the messages. In the original code, this would simply return concatenated
lines of strings, but since we're using objects we must store them in a
data structure (in this case a deque). The choice for using deque is
due to the fact deque was already being used originally within functions of
chatserver.cpp.