Add Vagrant file and Chef Solo cookbook for provisioning
A virtual machine running Ubuntu 16.04 (xenial) can now be provisioned
after installing Vagrant and VirtualBox. Run ```vagrant up``` in the
project's directory to create a virtual machine configured to be able to
run cmake and make on the project. To connect to the virtual machine, enter
```vagrant ssh```. Files in ```/home/vagrant/adventure2019``` is a
shared directory with the local copy on your computer and changes will
be reflected accordingly.