Add basic login and registration
Added a 'main menu' for handling users who have not yet logged in. Users
can choose to register a new Player to the game or log into an existing
one with a valid username and password.

In addition, the Player class 'password' member variable type has been
changed to std::string.