Replace GameResponse class with Response struct in Game class
The GameResponse class has been removed due to being a container for
primitive data types. A struct called Response is defined in the Game
class which performs the same role.

Client IDs have been converted to uintptr_t type from unsigned long int
type, as this is the type defined in Server.h for Connected IDs.