Add combat checks to movement commands and start room check on loss
A player will not be able to move to a different room if they are in
combat. They will be given a message if they try. Admins who use the
`goto` command will exit their combat instance.

The `CombatHandler::exitCombat()` methods have been changed to public
methods for the purpose of exiting out of combat using the Admin role's
`goto` command.

On losing a battle, the player will be placed in the starting room
with ID `Player::STARTING_LOCATION`. If such room does not exist, they
will be placed in the first room of the first area in the world.